sessions/session4/aws/requirements.txt
pandas
psycopg2-binary
pytest
pytest-cov
python-dotenv
boto3
openpyxl
dbt-postgres==1.7.4